extract fragment file
