+++ apps/web/src/pages/Users/AdminUserProfilePage/AdminUserProfilePage.tsx
@@ -17,7 +17,7 @@ import useRequiredParams from "~/hooks/useRequiredParams";
 import RequireAuth from "~/components/RequireAuth/RequireAuth";
 import useUserDownloads from "~/hooks/useUserDownloads";
 import DownloadDocxButton from "~/components/DownloadButton/DownloadDocxButton";
-import { FlexibleWorkLocationOptions_Fragment } from "~/components/Profile/components/WorkPreferences/Display";
+import { FlexibleWorkLocationOptions_Fragment } from "~/components/Profile/components/WorkPreferences/fragment";
 
 const AdminUserProfileUser_Fragment = graphql(/* GraphQL */ `
   fragment AdminUserProfileUser on User {
